Move SDN logic to separate package
