+++ pkg/transform/sdn/testdata/expected-network-cr-master.yaml
